Move all tests into base package
