+++ src/test/java/io/github/incplusplus/peerprocessing/ClientIT.java
@@ -1,9 +1,10 @@
-package io.github.incplusplus.peerprocessing.integration.client;
+package io.github.incplusplus.peerprocessing;
 
 import io.github.incplusplus.peerprocessing.client.Client;
 import io.github.incplusplus.peerprocessing.server.Server;
 import io.github.incplusplus.peerprocessing.slave.Slave;
-import org.junit.jupiter.api.*;
+import org.junit.jupiter.api.AfterAll;
+import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvFileSource;
 
@@ -14,24 +15,20 @@ import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 import java.util.concurrent.FutureTask;
 
-import static io.github.incplusplus.peerprocessing.integration.client.ClientIT.VERBOSE_TEST_OUTPUT;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-/**
- * Tests in this class are "jerky" because they involve the server
- * starting up and shutting down between tests.
- */
-class JerkyClientIT {
-	private int serverPort;
+public class ClientIT {
+	public static boolean VERBOSE_TEST_OUTPUT = false;
+	private static int serverPort;
 	
-	@BeforeEach
-	void setUp() throws IOException {
+	@BeforeAll
+	static void setUp() throws IOException {
 		serverPort = Server.start(0, VERBOSE_TEST_OUTPUT);
 		while (!Server.started()) {}
 	}
 	
-	@AfterEach
-	void tearDown() throws IOException {
+	@AfterAll
+	static void tearDown() throws IOException {
 		Server.stop();
 	}
